pyproject.toml
[tool.poetry]
name = "photonumbergamebot"
version = "0.1.0"
description = "Telegram bot based on aiogram for a fun photo number game"
authors = ["thunderpeal"]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.15.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
